requirements.txt
hubspot-api-python
requests